Capture works!
The program uses an auxiliary program that does the actual capture -- we authenticate, then run that program with enough privilege to read /dev/bpf*. That program writes the capture to a temporary file, which the main program then reads. I do this, to support not running the entire program as root, nor by needing the /dev/bpf* devices readable by the user, and just use OS X authentication to access the packet filter.

Also, the filter list is a "dynamic" list, and doesn't use a table view. The capture panel resizes to accommodate the list of filter predicates, like the "rules" list in Mail.app. It's slick, but limited to 15 predicates. I've removed the parts written by others that implemented the custom-view table view.